Authentication
Build a login service and fulfill the below User stories.

User story 1: Create a homepage where the user can either log in or move to the new account webpage

User story 2: A webpage that shows up when the user has tried to log in but entered an illegal name or password.

User story 3: A “create account” webpage that allows the user to create a new account. Please check if a user with the same email id already exists, don't allow for registration.

User story 4: Display the name of the user and an option for logout on the dashboard page after log in.